add tool bar
