+++ app/src/main/java/com/example/frimo/fragment/Everytime_FRIMO.java
@@ -1,4 +1,4 @@
-package com.example.frimo;
+package com.example.frimo.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,14 +7,16 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
-public class Little_Me_Diary extends Fragment {
+import com.example.frimo.R;
+
+public class Everytime_FRIMO extends Fragment {
 
     // Fragment를 만들고 레이아웃이랑 연결
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.little_me_diary, container, false);
+                R.layout.everytime_frimo, container, false);
 
         return rootView;
     }
