+++ app/src/main/java/com/example/frimo/fragment/Trend_Report.java
@@ -1,4 +1,4 @@
-package com.example.frimo;
+package com.example.frimo.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,14 +7,16 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
-public class Friendly_Community extends Fragment {
+import com.example.frimo.R;
+
+public class Trend_Report extends Fragment {
 
     // Fragment를 만들고 레이아웃이랑 연결
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.friendly_community, container, false);
+                R.layout.trend_report, container, false);
 
         return rootView;
     }
