+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,8 @@
 
     <androidx.appcompat.widget.Toolbar
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:background="@color/purple_200">
 
         <androidx.appcompat.widget.AppCompatImageView
             android:layout_width="35dp"
